Initial import of amqpprox.
This commit collapses the git history of the repository and extracts the
sections of our internal amqpprox repository that are intended for external
consumption. Internally the full history up until this point is preserved under
the 'pre-history' branch.

Although collapsing the git history to a single commit, we do not want
to lose any recognition for those that have contributed to amqpprox up
until this point. Those people are recorded in `docs/thanks.md` and we
thank them for their contributions.