Send amqpprox host info to RabbitMQ in StartOk
Add a new field to StartOk ConnectionProperties called 'amqpprox_host'
which is populated by the gethostname() of the machine, plus the port
number of the local end of the TCP socket. This should be enough
information to tie together the proxy host/instance of connections
shown on RabbitMQ Management UI/logs.

I tested this out by running amqpprox on my machine & connecting to it
locally w/ a remote backend setup (direct to a rabbitmq broker).
The amqpprox_host field shows up with the correct port, verified with Wireshark

hostname is resolved at startup & the value is passed down into the
Session objects are they're created to facilitate testing. I didn't
think we needed to worry about gethostname returning different results
at runtime - on windows at least this requires a machine restart.