PacketProcessor: Stop assuming full protocol header is in one read
Noticed while testing TLS that it's common for the AMQP 0.9.1 header
to be split across multiple reads. Although this is very surprising
since it's only 8 bytes, from testing on my machine it feels fairly common.

A splunk search for 'Incorrect header passed' also shows thousands of
results in the last 30 days, though I'm unclear if this is length
or content related.

This change uses the same data recycling technique as for frame
buffering, but for the header.