Fix multiple socket listeners
While testing we noticed that listening on multiple sockets wasn't working properly,
typically the second socket to listen was the only one to work.

It turns out this was because the ssl context held by d_socket was destructed as the
second listener's doAccept method was called. This context is passed by reference into the
asio socket, and also used by refreshSocketContext.

d_socket is only held by Server to keep it alive between doAccept() and the
async_accept callback being invoked. Changing this to a shared_ptr and passing
ownership into the callback does the same thing