perf tester: Handle channel close, extend listen backlog
Noticed during testing that amiquip was sending Channel Close but the server
send Channel OpenOk in return, triggering an error.

Since this happened while the connection was being Drop'ed the error seems
to have been swallowed.

The connection close is now explicit, even though this does avoid closing
the channel. I left in the channel close -> closeok in case we change something
here in future.