Automatically reconnect paused connections held at Connection.Open
During unpause it's important we disconnect any client which is
connected to a broker, because the broker mapping might be changed
during the pause duration. But we can be nicer to connections which
are held just after sending Connection.Open to amqpprox. This should
let us reduce the vhost migration flow to just one disconnection
noticed by clients.