Make MaybeSecureSocketAdapter::isSecure() private
This is only used for something which could happen upfront, so I moved
some things around to make this the case.

I made this change when I thought we'd want to template this class on
whether it was tls or not. Since beast::basic_stream isn't going to work
for us it isn't necessary anymore, but I thought it was still useful