Extend socket and timer lifetimes beyond data rate timer handler's
It turns out that although asio does cancel timers as they are being destructed,
the handlers only run after the timer and socket has been destructed. This has two
issues:

1. The timer object itself is destructed, which (according to valgrind at least, I
couldn't find anything in the asio docs about this) seems to be bad.
2. The handler was unconditionally using the MaybeSecureSocketAdaptor this pointer
which was clearly bad.

This has the potential to cause some pretty nasty issues when using data rate limits