Ensure partially read frame buffer is always large enough to hold frame
amqpprox internally buffers partial frames until the full frame is in memory,
at which point they are sent to the broker or client (depending on direction).

If there is no data internally buffered when data becomes available on the socket,
amqpprox creates a buffer of exactly the required size. Sometimes - e.g. When amqpprox
is quite busy - the requested buffer is larger than expected. This causes the
partial-frame buffer swap operation to fail.